Develop

+++ 3d-virtual-nlu-api/src/main/java/vn/edu/hcmuaf/virtualnluapi/entity/EmailVerification.java
@@ -13,8 +13,7 @@ import java.sql.Timestamp;
 @Builder
 public class EmailVerification {
     private int id;
-    private long userId;
+    private int userId;
     private String token;
     private Timestamp expiredAt;
-
 }
